Removed internal namespace and cleaned up.
