+++ ts/Extensions/Annotations/Controllables/ControllableCircle.ts
@@ -142,7 +142,7 @@ class ControllableCircle extends Controllable {
 
 /* *
  *
- *  Class Properties
+ *  Class Prototype
  *
  * */
